:bug: #50 added dedicated AppScreen with ViewModel
