+++ shared/src/commonMain/kotlin/App.kt
@@ -1,10 +1,4 @@
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import co.touchlab.kermit.Logger
-import dev.icerock.moko.mvvm.viewmodel.ViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
 import net.schacher.mcc.shared.auth.AuthHandler
 import net.schacher.mcc.shared.auth.PersistingAuthHandler
 import net.schacher.mcc.shared.datasource.database.CardDatabaseDao
@@ -19,8 +13,8 @@ import net.schacher.mcc.shared.platform.platformModule
 import net.schacher.mcc.shared.repositories.CardRepository
 import net.schacher.mcc.shared.repositories.DeckRepository
 import net.schacher.mcc.shared.repositories.PackRepository
-import net.schacher.mcc.shared.screens.login.LoginScreen
-import net.schacher.mcc.shared.screens.main.MainScreen
+import net.schacher.mcc.shared.screens.app.AppScreen
+import net.schacher.mcc.shared.screens.app.AppViewModel
 import net.schacher.mcc.shared.screens.main.MainViewModel
 import net.schacher.mcc.shared.screens.mydecks.MyDecksViewModel
 import net.schacher.mcc.shared.screens.newdeck.NewDeckViewModel
@@ -28,9 +22,7 @@ import net.schacher.mcc.shared.screens.packselection.PackSelectionViewModel
 import net.schacher.mcc.shared.screens.search.SearchViewModel
 import net.schacher.mcc.shared.screens.settings.SettingsViewModel
 import net.schacher.mcc.shared.screens.spotlight.SpotlightViewModel
-import net.schacher.mcc.shared.utils.debug
 import org.koin.compose.KoinApplication
-import org.koin.compose.koinInject
 import org.koin.core.KoinApplication
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
@@ -57,9 +49,7 @@ val viewModels = module {
 }
 
 @Composable
-fun App(
-    databaseDao: DatabaseDao, onKoinStart: KoinApplication.() -> Unit = {}
-) {
+fun App(databaseDao: DatabaseDao, onKoinStart: KoinApplication.() -> Unit = {}) {
     val authHandler = PersistingAuthHandler(databaseDao as SettingsDao)
     KoinApplication(application = {
         onKoinStart()
@@ -79,36 +69,5 @@ fun App(
     }
 }
 
-@Composable
-fun AppScreen(appViewModel: AppViewModel = koinInject()) {
-    val loggedIn = appViewModel.state.collectAsState()
-    if (loggedIn.value) {
-        MainScreen()
-    } else {
-        LoginScreen {
-            appViewModel.onGuestLoginClicked()
-        }
-    }
-}
 
-class AppViewModel(
-    private val authHandler: AuthHandler
-) : ViewModel() {
 
-    private val _state = MutableStateFlow(authHandler.isLoggedIn())
-
-    val state = _state.asStateFlow()
-
-    init {
-        viewModelScope.launch {
-            authHandler.loginState.collect {
-                Logger.debug { "Login state changed: $it" }
-                _state.value = it
-            }
-        }
-    }
-
-    fun onGuestLoginClicked() {
-        _state.value = true
-    }
-}
